Move canvas to its own file
