+++ src/main/kotlin/edu/cwru/raycaster/ContextualCanvas.kt
@@ -1,0 +1,47 @@
+package edu.cwru.raycaster
+
+import javafx.scene.canvas.Canvas
+import javafx.scene.canvas.GraphicsContext
+import javafx.scene.image.Image
+import javafx.scene.paint.Color
+import javafx.scene.paint.Paint
+
+class ContextualCanvas(private val width: Int, private val height: Int): Canvas(width.toDouble(), height.toDouble()) {
+    val context: GraphicsContext = graphicsContext2D
+    var fill: Paint
+        get() = context.fill
+        set(color) {
+            context.fill = color
+        }
+    var stroke: Paint
+        get() = context.stroke
+        set(color) {
+            context.stroke = color
+        }
+
+    fun drawImage(image: Image, x: Number, y: Number) {
+        context.drawImage(image, x.toDouble(), y.toDouble())
+    }
+
+    fun fillRect(x: Number, y: Number, w: Number, h: Number, color: Color? = null) {
+        val oldFill = fill
+        if (color != null) {
+            fill = color
+        }
+        context.fillRect(x.toDouble(), y.toDouble(), w.toDouble(), h.toDouble())
+        fill = oldFill
+    }
+
+    fun strokeLine(x1: Number, y1: Number, x2: Number, y2: Number, color: Color? = null) {
+        val oldStroke = stroke
+        if (color != null) {
+            stroke = color
+        }
+        context.strokeLine(x1.toDouble(), y1.toDouble(), x2.toDouble(), y2.toDouble())
+        stroke = oldStroke
+    }
+    fun strokeLine(p1: Vec2Double, p2: Vec2Double, color: Color? = null) =
+        strokeLine(p1.x, p1.y, p2.x, p2.y, color)
+    fun strokeLine(p1: Vec2Int, p2: Vec2Int, color: Color? = null) =
+        strokeLine(p1.x, p1.y, p2.x, p2.y, color)
+}
